+++ .github/workflows/merge.yml
@@ -71,7 +71,7 @@ jobs:
             verification_path: actuator/health
             file: backend/openshift.deploy.yml
             parameters:
-              -p RESULTS_ENV_OPENSEARCH=test
+              -p SILVA_ENV_OPENSEARCH=test
           - name: frontend
             file: frontend/openshift.deploy.yml
             parameters:
@@ -147,7 +147,7 @@ jobs:
             verification_path: actuator/health
             file: backend/openshift.deploy.yml
             parameters:
-              -p RESULTS_ENV_OPENSEARCH=production
+              -p SILVA_ENV_OPENSEARCH=production
           - name: frontend
             file: frontend/openshift.deploy.yml
             parameters:
