+++ backend/src/main/java/ca/bc/gov/restapi/silva/ResultsApplication.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.restapi.results;
+package ca.bc.gov.restapi.silva;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
