+++ backend/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 # General application
-logging.level.ca.bc.gov.restapi.results = ${LOGGING_LEVEL:INFO}
-spring.application.name = results-api
+logging.level.ca.bc.gov.restapi.silva = ${LOGGING_LEVEL:INFO}
+spring.application.name = silva-api
 server.error.include-message=always
 server.port = ${SERVER_PORT:8080}
 
@@ -27,7 +27,7 @@ spring.datasource.hikari.connectionTimeout = ${DB_POOL_CONN_TIMEOUT:90000}
 spring.datasource.hikari.idleTimeout = ${DB_POOL_IDLE_TIMEOUT:45000}
 spring.datasource.hikari.maxLifetime = ${DB_POOL_MAX_LIFETIME:30000}
 spring.datasource.hikari.keepaliveTime = 30000
-spring.datasource.hikari.poolName = NrResultsDbPool
+spring.datasource.hikari.poolName = NrSilvaDbPool
 spring.datasource.hikari.minimumIdle = 1
 spring.datasource.hikari.maximumPoolSize = 3
 spring.jpa.properties.hibernate.default_schema = silva
@@ -35,6 +35,6 @@ spring.jpa.database-platform = org.hibernate.dialect.OracleDialect
 spring.jpa.show-sql = true
 
 # OpenSearch settings
-nr-results-ecs-version = 8.9
-nr-results-backend-env-opensearch = ${RESULTS_ENV_OPENSEARCH:development}
-nr-results-team-email-address = Team.Silva@gov.bc.ca
+nr-silva-ecs-version = 8.9
+nr-silva-backend-env-opensearch = ${SILVA_ENV_OPENSEARCH:development}
+nr-silva-team-email-address = Team.Silva@gov.bc.ca
