+++ backend/src/test/java/ca/bc/gov/restapi/silva/ResultsApplicationTests.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.restapi.results;
+package ca.bc.gov.restapi.silva;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
